requirements.txt
attrs==16.3.0
cffi==1.9.1
click==6.6
constantly==15.1.0
cryptography==1.6
cssselect==1.0.0
decorator==4.0.10
dj-database-url==0.4.1
Django==1.10.4
flake8==3.2.1
httplib2==0.9.2
hubstorage==0.23.6
idna==2.1
incremental==16.10.1
ipython==5.1.0
ipython-genutils==0.1.0
lxml==3.6.4
mccabe==0.5.2
parsel==1.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycodestyle==2.2.0
pycparser==2.17
PyDispatcher==2.0.5
pyflakes==1.3.0
Pygments==2.1.3
pyOpenSSL==16.2.0
python-dotenv==0.6.1
PyYAML==3.12
queuelib==1.4.2
recordtype==1.1
requests==2.12.3
retrying==1.3.3
scrapinghub==1.9.0
Scrapy==1.2.1
service-identity==16.0.0
shub==2.4.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Twisted==16.6.0
w3lib==1.16.0
wcwidth==0.1.7
zope.interface==4.3.2
